refactor(no-test-callback): rename rule to `no-done-callback` (#653)
BREAKING CHANGE: rename `no-test-callback` to `no-done-callback`
